Conflicts:
	com.getshop.client/apps/InvoicePayment/InvoicePayment.php
	com.getshop.client/apps/InvoicePayment/template/config.phtml
	core/src/main/java/com/thundashop/core/pdf/InvoiceGenerator.java
	core/src/main/java/com/thundashop/core/pdf/InvoiceManager.java
	messages/src/main/java/com/thundashop/core/pdf/data/AccountingDetails.java

+++ messages/src/main/java/com/thundashop/core/c3/C3UserProjectPeriode.java
@@ -11,6 +11,6 @@ package com.thundashop.core.c3;
  */
 public class C3UserProjectPeriode extends ProjectCost {
     public String costType = "userperiode";
-    public int percent;
+    public double percent;
     public boolean nfr;
 }
